+++ src/codemagic/google/resource_managers/__init__.py
@@ -1,0 +1,1 @@
+from .release_manager import ReleaseManager