+++ src/codemagic/tools/firebase_app_distribution/action_groups/__init__.py
@@ -1,0 +1,1 @@
+from .releases_action_group import ReleasesActionGroup